refactor(core): move the workflow registries into @bakin/core/workflows
git-mv source-registry, node-type-registry, notification-channel-registry,
and agent-package-skill-registry from plugins/workflows/lib/ to
packages/core/src/workflows/ — bodies byte-for-byte, only import
specifiers adjusted. These registries are core plugin-loader
infrastructure (ctx.registerWorkflow / registerNodeType /
registerNotificationChannel and agent-package loading all flow through
them), so core importing them from one plugin's source tree inverted the
adapter/plugin boundary (audit WS2 finding). All core, plugin, and test
consumers repoint to @bakin/core/workflows/*; registration behavior is
unchanged.
