+++ plugins/workflows/lib/search-sync.ts
@@ -14,8 +14,8 @@ import { join } from 'path'
 import yaml from 'js-yaml'
 import type { PluginContext } from '@bakin/core/plugin-types'
 import type { WorkflowDefinition, WorkflowInstance } from '../types'
-import type { DefinitionSource } from './source-registry'
-import { getManagedDefinition } from './source-registry'
+import type { DefinitionSource } from '@bakin/core/workflows/source-registry'
+import { getManagedDefinition } from '@bakin/core/workflows/source-registry'
 import { loadDefinition, listDefinitions } from './parser'
 import { loadInstance } from './runtime'
 import { isWorkflowDisabled } from './availability'
