+++ src/core/plugin-host/reload-pipeline.ts
@@ -39,8 +39,8 @@ import { pluginRegistry } from '@/core/plugin-registry'
 import { removePluginSkillsByPlugin } from '@bakin/core/skills/plugin-skill-registry'
 import { getHookRegistry } from '@bakin/core/hooks/hook-registry-singleton'
 import { removeExecToolsByPlugin } from '../exec-tools/registry'
-import { unregisterPluginNodeTypes } from '../../../plugins/workflows/lib/node-type-registry'
-import { unregisterPluginNotificationChannels } from '../../../plugins/workflows/lib/notification-channel-registry'
+import { unregisterPluginNodeTypes } from '@bakin/core/workflows/node-type-registry'
+import { unregisterPluginNotificationChannels } from '@bakin/core/workflows/notification-channel-registry'
 import { unregisterPluginHealthChecks } from '../health-check-registry'
 import { ensurePluginSearchReady, unregisterContentTypesByPlugin } from '@/core/search-registry'
 import {
